+++ src/mantle/testing/exceptions/class-response-exception.php
@@ -18,8 +18,9 @@ abstract class Response_Exception extends Exception {
 	 *
 	 * @param int         $status  The HTTP status code.
 	 * @param string|null $message Optional exception message.
+	 * @param int         $code    Optional exception code.
 	 */
-	public function __construct( public readonly int $status, ?string $message = null ) {
-		parent::__construct( $message ?? "HTTP Response with status {$status}", $status );
+	public function __construct( public readonly int $status, ?string $message = null, int $code = 0 ) {
+		parent::__construct( $message ?? "HTTP Response with status {$status}", $code );
 	}
 }